add confirm dialog

+++ src/fe/motel/src/app/pages/service/service.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { ToastrModule, ToastrService } from 'ngx-toastr';
+import { FormatCurrency } from 'src/app/common/stringFormat';
 import { AppService } from 'src/app/services/app.service';
-
+declare var bootbox:any;
 @Component({
   selector: 'app-service',
   templateUrl: './service.component.html',
@@ -12,13 +13,15 @@ export class ServiceComponent implements OnInit {
   provides = [];
   provideInsert = {
     name: "",
-    type: 0
+    type: 0,
+    defaultPrice: 0
   }
 
   provideUpdate = {
     id :"",
     type: 0,
-    name: ""
+    name: "",
+    defaultPrice : 0
   }
   constructor(
     private _service : AppService,
@@ -49,14 +52,31 @@ export class ServiceComponent implements OnInit {
   }
 
   removeProvide(id){
-    this._service.deleteProvide(id).subscribe(
-      response =>{
-        if(response.isSucceeded == true){
-          this._toast.success("Đã xóa dịch vụ");
-          this.getProvides();
-        }
+    bootbox.confirm({
+      title: "Bạn chắc chắn muốn xóa dịch vụ này?",
+      message: "Khi dịch vụ được xóa sẽ mất dữ liệu ở các khu trọ sử dụng dịch vụ",
+      buttons: {
+          cancel: {
+              label: '<i class="fa fa-times"></i> Hủy bỏ'
+          },
+          confirm: {
+              label: '<i class="fa fa-check"></i> Xác nhận'
+          }
+      },
+      callback: function (result) {
+          if(result){
+            this._service.deleteProvide(id).subscribe(
+              response =>{
+                if(response.isSucceeded == true){
+                  this._toast.success("Đã xóa dịch vụ");
+                  this.getProvides();
+                }
+              }
+            )
+          }
       }
-    )
+  });
+    
   }
 
   getProvideUpdate(provide){
@@ -72,6 +92,9 @@ export class ServiceComponent implements OnInit {
         }
       }
     )
+  }
+  formatCurrency(input){
+    return FormatCurrency(input);
   }
 
 }
